# server/docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Platform for event management: authentication, user profiles, events with comments and registration",
        "title": "Eventor API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://eventor.local/support",
            "email": "support@eventor.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/events": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get paginated list of events with optional search filter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "List events",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page (1-100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by title or description",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of events",
                        "schema": {
                            "$ref": "#/definitions/internal_event_transport_http_controllers_event.ListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/internal_event_transport_http_controllers_event.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create an event by authenticated user. Only creator can modify/delete this event later.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Create new event",
                "parameters": [
                    {
                        "description": "Event creation data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/eventor_internal_event_services_usecase_event.CreateEventRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Event created successfully",
                        "schema": {
                            "$ref": "#/definitions/eventor_internal_event_transport_http_dto_event.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation failed",
                        "schema": {
                            "$ref": "#/definitions/internal_event_transport_http_controllers_event.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/internal_event_transport_http_controllers_event.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden: insufficient permissions",
                        "schema": {
                            "$ref": "#/definitions/internal_event_transport_http_controllers_event.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_event_transport_http_controllers_event.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/events/image": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload image file for event cover. Returns storage key and URL. Accepts multipart/form-data.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Upload event cover image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image file (JPEG, PNG, WebP)",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image uploaded successfully",
                        "schema": {
                            "$ref": "#/definitions/internal_event_transport_http_controllers_event.EventImageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid file or not an image",
                        "schema": {
                            "$ref": "#/definitions/internal_event_transport_http_controllers_event.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_event_transport_http_controllers_event.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "File storage error",
                        "schema": {
                            "$ref": "#/definitions/internal_event_transport_http_controllers_event.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/events/images/{image_id}/file": {
            "get": {
                "description": "Download event cover image by image ID. Returns raw image bytes.",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get event image file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image ID",
                        "name": "image_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Invalid image ID",
                        "schema": {
                            "$ref": "#/definitions/internal_event_transport_http_controllers_event.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Image not found",
                        "schema": {
                            "$ref": "#/definitions/internal_event_transport_http_controllers_event.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Storage error",
                        "schema": {
                            "$ref": "#/definitions/internal_event_transport_http_controllers_event.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_event_transport_http_controllers_event.ErrorResponse": {
            "description": "Standard error response format",
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code\n@example 404",
                    "type": "integer"
                },
                "details": {
                    "description": "Detailed validation errors (if applicable)\n@example {\"title\": \"min length is 3\"}",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "error": {
                    "description": "Human-readable error message\n@example \"event not found\"",
                    "type": "string"
                }
            }
        },
        "internal_event_transport_http_controllers_event.EventImageResponse": {
            "description": "Response after successful image upload",
            "type": "object",
            "properties": {
                "image_id": {
                    "description": "Unique image identifier in storage\n@example \"img_abc123xyz\"",
                    "type": "string"
                },
                "url": {
                    "description": "URL to retrieve the image file\n@example \"/api/v1/events/images/img_abc123xyz/file\"",
                    "type": "string"
                }
            }
        },
        "internal_event_transport_http_controllers_event.ListResponse": {
            "description": "Paginated list of events",
            "type": "object",
            "properties": {
                "count": {
                    "description": "Number of events in current page\n@example 20",
                    "type": "integer"
                },
                "events": {
                    "description": "List of events",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/eventor_internal_event_transport_http_dto_event.EventResponse"
                    }
                },
                "limit": {
                    "description": "Requested page size\n@example 20",
                    "type": "integer"
                },
                "offset": {
                    "description": "Requested offset\n@example 0",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter JWT token in format: `Bearer \u003cyour_token_here\u003e`. Token obtained from /auth/login or /auth/register.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}